backend/requirements.txt
flask
flask-cors
flask-dance[github]
requests
python-dotenv
psycopg2-binary